refactor: update logic
